Add files via upload
IMPLEMENTATION OF KNN
ALGORITHM TO IRIS DATABASE 

Using matlab, a developed code from scratch of KNN algorithm was created.

The data chosen was a selection of 15 points, from each class label given. 15 points from ‘setosa’, 15 points from ‘versicolor’, and 15 points from ‘virginica’. This way we get an equal amount of training/test points with little bias.
From here, we choose the first 2 columns, sepal length and sepal width, to be our X and Y. This
goes for training and testing data. We calculate distance, using the Euclidean distance formula.
Next, the distances are sorted for each column, with the smallest value being moved to the top
of the data, in each column. From here, we can get the original position of the distance, using
dis_labels before it is sorted. We can then get the label of the value of the smallest distance
and use that for our prediction. Since we choose the first 15 values to be setosa, the next 15 to be versicolor, and the last 15 to be virginica, we can assign the numeric label value to the label name. For example a distance chosen could be from original line 32, being a part of the virginica label. Finally, we assign the 3 labels a numeric value, from 1 to 3, 1 being setosa, 2 being versicolor and 3 being virginica. This allows us to take the mode of each column, giving us our output. Since we choose k to be equal to 3, We only take the 3 shortest distances, of our available 45 rows. From here, the mode of each column is taken, giving us the output, pred_labels.
In code  we must randomize the data over 100 iterations, for k varying from 1 to 8. We set up 2 for loops with our randomization code being set up immediately. This time, we must take note of index_labels, being the original position of the randomized data, we must know where the original data was classified as, in order to get a proper estimation. The data is then split , in a train/test ratio of 70/30. From here, the process is the same as part A, calculate
the distance, and sort the results. Again, we take the mode of each column, and the resulting row is then placed into a matrix, predicted_label. In the first K loop, it will only have 1 row to take the mode of, but as the loop continues, it has up to 8 rows of data to consider for the mode. Finally, the accuracy_matrix is set up with dimensions 100x8, matching our 100 iterations over 8 K values.
The average is then taken over each column, giving us the average accuracy for each K.
Standard deviation is calculated from the original accuracy matrix and the plot can be seen
here.